+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/wand/WandUseMode.java
@@ -1,0 +1,5 @@
+package com.elmakers.mine.bukkit.api.wand;
+
+public enum WandUseMode {
+    SUCCESS, PRECAST, ALWAYS
+}